continuing to improve the visualization
