+++ Cargo.toml
@@ -14,7 +14,7 @@ keywords = ["opensearch"]
 
 [features]
 default = ["visualizer"]
-visualizer = ["dep:askama", "dep:uuid"]
+visualizer = ["dep:askama"]
 
 [dependencies]
 serde = { version = "1.0.219", features = ["derive"] }
@@ -22,4 +22,4 @@ serde_json = "1.0.140"
 
 # Optional visualization dependencies
 askama = { version = "0.14.0", optional = true }
-uuid = { version = "1.18.0", optional = true, features = ["v4"] }
+# uuid = { version = "1.18.0", optional = true, features = ["v4"] }
